+++ utility/5x3.cpp
@@ -38,28 +38,28 @@ const size_t enable_pins[CHANNEL_COUNT] =
 // Units
 
 // Properties
-const double steps_per_position_units_default[CHANNEL_COUNT] =
+const long steps_per_position_units_default[CHANNEL_COUNT] =
   {
     steps_per_position_units_element_default,
     steps_per_position_units_element_default,
     steps_per_position_units_element_default,
   };
 
-const double velocity_max_default[CHANNEL_COUNT] =
+const long velocity_max_default[CHANNEL_COUNT] =
   {
     velocity_max_element_default,
     velocity_max_element_default,
     velocity_max_element_default,
   };
 
-const double velocity_min_default[CHANNEL_COUNT] =
+const long velocity_min_default[CHANNEL_COUNT] =
   {
     velocity_min_element_default,
     velocity_min_element_default,
     velocity_min_element_default,
   };
 
-const double acceleration_max_default[CHANNEL_COUNT] =
+const long acceleration_max_default[CHANNEL_COUNT] =
   {
     acceleration_max_element_default,
     acceleration_max_element_default,
@@ -108,7 +108,7 @@ const bool switch_soft_stop_enabled_default[CHANNEL_COUNT] =
     false,
   };
 
-const double home_velocity_default[CHANNEL_COUNT] =
+const long home_velocity_default[CHANNEL_COUNT] =
   {
     home_velocity_element_default,
     home_velocity_element_default,
